MAIN/server/requirements.txt
fastapi[standard]
uvicorn
pillow
numpy